Rename herd-vote to quiz across app
