--- src/app/[lang]/apps/herd-vote/page.tsx
@@ -1,16 +1,0 @@
-import { redirect } from 'next/navigation'
-import { normalizeUrlLocale } from '@/lib/i18n-routing'
-import { type Locale, SUPPORTED_LOCALES } from '@/i18n/config'
-import { notFound } from 'next/navigation'
-
-// Merge Herd Vote route into quiz page
-export const dynamic = 'force-dynamic'
-
-type P = { params: Promise<{ lang: string }> }
-
-export default async function HerdVoteRedirect({ params }: P) {
-  const { lang: raw } = await params
-  const lang = normalizeUrlLocale(raw)
-  if (!SUPPORTED_LOCALES.includes(lang as Locale)) notFound()
-  redirect(`/${lang}/apps/quiz`)
-}
