+++ src/app/[lang]/play/[code]/_Join.tsx
@@ -64,7 +64,7 @@ export default function PlayJoin({ code: passedCode }: { code?: string }) {
 
   return (
     <div className="mx-auto max-w-xl p-6 space-y-4">
-      <h1 className="text-2xl font-bold">Herd Vote – pripojenie</h1>
+      <h1 className="text-2xl font-bold">Quiz – pripojenie</h1>
       <p className="text-sm text-gray-600">
         Naskenujte QR kód od moderátora alebo použite zdieľaný odkaz. Zadajte svoje meno a
         počkajte na spustenie hry.
