+++ src/app/api/games/[code]/rounds/results/route.ts
@@ -3,7 +3,7 @@ import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
 import { RealtimeServer } from '@/lib/realtime/server'
 import { channelFor } from '@/lib/realtime/types'
-import { calculateRoundScores, mapFromGameScoringMode } from '@/lib/herdvote/scoring'
+import { calculateRoundScores, mapFromGameScoringMode } from '@/lib/quiz/scoring'
 import { supabaseServer } from '@/integrations/supabase/server'
 import { getSession } from '@/app/api/games/_session'
 import { asArray } from '@/lib/supabase/safe'
