+++ src/app/api/quiz/categories/route.ts
